generate_config.py: assembly CONFIG_SFP%02u_PARAMS before write to file
When output of a print is redirected to a file it adds a newline. To
avoid this, the comma at the end was added. However, this solution adds
a space after the redirected text. To avoid this the line for
CONFIG_SFP%02u_PARAMS is assebled first, then written to a file.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>